refactor(RevenueCatUI): extract RevenueCatDialogScaffold from PaywallDialog
Move the dialog scaffold composable and its sizing helpers into a
reusable `RevenueCatDialogScaffold` in the composables package so it
can be shared by other dialog-based UI surfaces.
